requirements.txt
aiohttp==3.6.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
base-api-connector==0.3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==3.0.7
django-cors-headers==3.1.1
djangorestframework==3.10.3
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
mysqlclient==1.4.4
pycodestyle==2.5.0
pystache==0.5.4
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
trello==0.9.4
urllib3==1.24.2
Werkzeug==0.16.0
whitenoise==4.1.4
yarl==1.3.0
